+++ lib/hello_web/controllers/up_controller.ex
@@ -1,0 +1,17 @@
+defmodule HelloWeb.UpController do
+  use HelloWeb, :controller
+
+  def index(conn, _params) do
+    conn
+    |> put_resp_content_type("text/html")
+    |> send_resp(:ok, "")
+  end
+
+  def databases(conn, _params) do
+    Ecto.Adapters.SQL.query!(Hello.Repo, "SELECT 1")
+
+    conn
+    |> put_resp_content_type("text/html")
+    |> send_resp(:ok, "")
+  end
+end