+++ test/hello_web/controllers/page_controller_test.exs
@@ -5,9 +5,4 @@ defmodule HelloWeb.PageControllerTest do
     conn = get(conn, Routes.page_path(conn, :home))
     assert html_response(conn, 200)
   end
-
-  test "up page loads", %{conn: conn} do
-    conn = get(conn, Routes.page_path(conn, :up))
-    assert html_response(conn, 200)
-  end
 end